requirements.txt
Django==1.11.7
django-haystack==2.6.1
jieba==0.39
Markdown==2.6.10
psycopg2==2.7.3.2
Pygments==2.2.0
pytz==2017.3
Whoosh==2.7.4
